% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{checks.input.posthoc.analysis}
\alias{checks.input.posthoc.analysis}
\title{Internal function, used for performing input checks for input in \code{mcleod.posterior.estimates.random.effect} and \code{mcleod.predictive.CI}}
\usage{
checks.input.posthoc.analysis(
  X,
  N,
  mcleod_res,
  offset_vec,
  is_Noise_Poisson,
  covariates,
  skip_checks_X_N
)
}
\arguments{
\item{X}{See description in original function}

\item{N}{See description in original function}

\item{mcleod_res}{See description in original function}

\item{offset_vec}{See description in original function}

\item{is_Noise_Poisson}{See description in original function}

\item{covariates}{See description in original function}

\item{skip_checks_X_N}{Should checks be skipped for X and N. This is relecent in \code{mcleod.predictive.CI}, where the value of X is not known yet.}
}
\value{
posterior mean for slope cofficients if available. Else returns the value 0.
}
\description{
Internal function, used for performing input checks for input in \code{mcleod.posterior.estimates.random.effect} and \code{mcleod.predictive.CI}
}
\keyword{internal}
