% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes}
\alias{codes}
\title{Procedure codes}
\format{a \code{data.frame} with 1 variable
\describe{
  \item{px10_spine}{(chr) Procedure codes from ICD-10 for spine procedures}
  \item{px10_pectus}{(chr) Procedure codes from ICD-10 for pectus procedures}
  \item{px9_spine}{(chr) Procedure codes from ICD-9 for spine procedures}
  \item{px9_pectus}{(chr) Procedure codes from ICD-9 for pectus procedures}
  \item{cpt_spine}{(chr) Procedure codes from CPT for spine procedures}
  \item{cpt_pectus}{(chr) Procedure codes from CPT for pectus procedures}
}}
\source{
StateCountyFips15jun17.Rda
}
\usage{
codes
}
\description{
This table contains the medication names Lisa McLeod's group are interested in.
}
\keyword{datasets}
