% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{compute_CI_curves_function}
\alias{compute_CI_curves_function}
\title{Function performs efficient computation of pointwise CIs, for a given confidence level}
\usage{
compute_CI_curves_function(
  bank_original,
  res_mcleod_data,
  rho_calibration_obj,
  median_curve,
  alpha.one.sided,
  verbose
)
}
\arguments{
\item{bank_original}{Cache object, produced by}

\item{res_mcleod_data}{result of \code{mcleod} for the test data}

\item{rho_calibration_obj}{object of type \code{"mcleod.CI.obj.rho"}}

\item{median_curve}{median curve for the data, generated by running \code{compute_medians_curve} on the test data}

\item{alpha.one.sided}{T1E for one-sided testing}

\item{verbose}{Should progress be printed to screen}
}
\value{
a list with the following entries: \code{q_star_LE} - grid of upper bound, by points in \code{theta_vec}; \code{q_star_GE} -  grid of lower bounds, by points in \code{theta_vec}; \code{bank} - updated cache object, to be placed in calling function. For \code{q_star_LE} and \code{q_star_GE} the names attribute gives the values of theta for which pointwise confidence intervals are computed.
}
\description{
Function performs efficient computation of pointwise CIs, for a given confidence level
}
\keyword{internal}
