% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{compute_medians_curve}
\alias{compute_medians_curve}
\title{Given a mcleod object, compute the posterior median for the random effect distribution}
\usage{
compute_medians_curve(mcleod_for_data, list_ind = NA)
}
\arguments{
\item{mcleod_for_data}{mcleod object}
}
\value{
vector of length = length(a.vec), with edge values of 0 and 1, giving the pointwise posterior medians across a.vec.
}
\description{
Medians are taken pointwise across a.vec
}
\keyword{internal}
