% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{init_mcleod_random_intercept_regression}
\alias{init_mcleod_random_intercept_regression}
\title{An internal function used for running a random normal intercept regression, in order to initialize the slope coefficients in the MCMC chain.}
\usage{
init_mcleod_random_intercept_regression(
  x,
  n,
  covariates,
  offset_p,
  family = "binomial"
)
}
\arguments{
\item{x}{Observed Counts/ successes for binomial samples}

\item{n}{Total number of draws for binomial sampeles. For Poisson regression, set this to \code{NULL}}

\item{covariates}{matrix of covatiates, row are samples, not intercept term needed}

\item{offset_p}{vector of offset values for the linear predictor. If none are specified by the user, pass a vector of zeroes.}

\item{family}{either 'binomial' or 'poisson'.}
}
\value{
a vector of length \code{ncol(covariates)}, giving the estimated slope coefficients, from a random normal intercept.
}
\description{
Function is called from \code{\link{mcleod}}, when paramter \code{beta_init} is set to \code{NULL}. Will attempt to run \code{GLMER} model first, with Binomial/Poisson samples and a random normmal intercept. If fails, will use fixed Binomial/Poisson regression.
}
\keyword{internal}
