% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_Functions.R
\name{inv.log.odds}
\alias{inv.log.odds}
\title{Compute probability from log-odds}
\usage{
inv.log.odds(theta)
}
\arguments{
\item{theta}{}
}
\value{
exp(theta) / (1 +exp(theta))
}
\description{
Compute probability from log-odds
}
\examples{

library(mcleod)
inv.log.odds(0)

}
