% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.CI.deconv.bank.constructor}
\alias{mcleod.CI.deconv.bank.constructor}
\title{Generate a cache for estimated CDFs of the mixing distributions, computed for worstcase (GE/LE) data generaitons}
\usage{
mcleod.CI.deconv.bank.constructor(
  N_vec = NULL,
  CI_param,
  Use_Existing_Permutations_From_Object = NULL
)
}
\arguments{
\item{N_vec}{The number of draws, for binomial observation. (when using Poisson samples, this sets the offset term, using offset = log(N) )}

\item{CI_param}{The CI parameters object used for the pipeline}

\item{Use_Existing_Permutations_From_Object}{an existing object generated using mcleod.estimate.CI. the cachec is instantiated using the CDF samples available in this object. It is up to the user to verify the the object uses the same CI parameters and N vec.}
}
\value{

}
\description{
The object is used to cache the estimated CDFs by value of (theta,q). The values median_curve_GE/median_curve_LE (returned inside the object) are double-nested lists by theta and then q, holding samples of the estimated CDFs (across mcleod's a.vec representation), by value of theta,q.
This object is then used throughout the pipeline to also hold CI_param.
}
\keyword{internal}
