% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.CI.deconv.bank.get_median_curves_for_worst_case_hypothesis}
\alias{mcleod.CI.deconv.bank.get_median_curves_for_worst_case_hypothesis}
\title{Title}
\usage{
mcleod.CI.deconv.bank.get_median_curves_for_worst_case_hypothesis(
  bank,
  ind_theta,
  ind_q,
  nr.curves,
  is_GE = T,
  do_serial = T
)
}
\arguments{
\item{bank}{}

\item{ind_theta}{}

\item{ind_q}{}

\item{nr.curves}{}

\item{is_GE}{}

\item{do_serial}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
