% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.CI.find.ai.by.theta.and.rho}
\alias{mcleod.CI.find.ai.by.theta.and.rho}
\title{Title}
\usage{
mcleod.CI.find.ai.by.theta.and.rho(res_mcleod_object, theta, rho, is_GE = T)
}
\arguments{
\item{res_mcleod_object}{}

\item{theta}{}

\item{rho}{}

\item{is_GE}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
