% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.CI.PV.at_point}
\alias{mcleod.CI.PV.at_point}
\title{Title}
\usage{
mcleod.CI.PV.at_point(
  bank,
  ind_theta,
  ind_q,
  CDF_value,
  a_index,
  alpha = 0.05,
  nr.perm = 200,
  do_check_vs_noiseless_case = T,
  do_check_vs_minimum_number_of_required_iterations = T,
  is_GE = F,
  do_serial = F
)
}
\arguments{
\item{bank}{}

\item{ind_theta}{}

\item{ind_q}{}

\item{CDF_value}{}

\item{a_index}{}

\item{alpha}{}

\item{nr.perm}{}

\item{do_check_vs_noiseless_case}{}

\item{do_check_vs_minimum_number_of_required_iterations}{}

\item{is_GE}{}

\item{do_serial}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
