% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.estimate.CI.single.theta}
\alias{mcleod.estimate.CI.single.theta}
\title{A wrapper function for mcleod.estimate.CI, used to estimate a two-sided confidence interval for the percentile of the mixing distribution at a given log-odds value.}
\usage{
mcleod.estimate.CI.single.theta(
  X,
  N,
  theta,
  CI_param = mcleod.CI.estimation.parameters(),
  ratio_holdout = 0.1,
  verbose = F
)
}
\arguments{
\item{X}{Same parameter as in \code{\link{mcleod.estimate.CI}}.}

\item{N}{Same parameter as in \code{\link{mcleod.estimate.CI}}.}

\item{theta}{The log-odds value for which the confidence interval will be built. For example, using a value of \code{theta=0} will return a confidence interval, in terms of CDF values, for mixing distribution at \code{theta=0}.}

\item{CI_param}{Same parameter as in \code{\link{mcleod.estimate.CI}}. This parameter can be used to change the confidence level.}

\item{ratio_holdout}{Same parameter as in \code{\link{mcleod.estimate.CI}}.}

\item{verbose}{Same parameter as in \code{\link{mcleod.estimate.CI}}.}
}
\value{
A vector with two entries name 'Lower' and 'Upper', denoting the two-sided CI, in terms of CDF values, for a given theta (log-odds value).
}
\description{
The function receives input parameters as defined for \code{\link{mcleod.estimate.CI}}, together with a single log-odds value (theta). The function returns a confidence interval, in terms of CDF values, for the mixing distribution at the required point.
}
\details{
This function is actually a "convenience wrapper" for \code{\link{mcleod.estimate.CI}}. It is used to make sure that the only point where CIs are computed is at log-odds=theta, and that calibration of rho is also done for this point alone.
}
\examples{
 # For full description of package model and workflow,
 # including this function, Type browseVignettes(package = 'mcleod') 
 # in the R console and check the package vignette
 
 ################################
 ## Example - computing CI for a single value of theta
 ################################
 #Generate data
 
 n = 500
 N = rep(20,n) 
 set.seed(1)
 p = inv.log.odds(rnorm(n,-2,0.5)+3*rbinom(n,1,0.3))
 X = rbinom(n = n,size = N,prob = p)
   
 CI.res = mcleod.estimate.CI.single.theta(X = X, N = N,
 
   # the quantile for which the CI is needed
   theta = 0,
   
   # resolution for the CI, in terms of CDF values
   CI_param = mcleod.CI.estimation.parameters(q_vec = seq(0.05,0.95,0.01),
                                              rho.set.value = 0.25) )
 
 CI.res
}
