% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.get.CIs.mixing.dist}
\alias{mcleod.get.CIs.mixing.dist}
\title{Extract pointwise confidence intervals for the CDF of the mixing distribution, from result of function \code{mcleod.estimate.CI}}
\usage{
mcleod.get.CIs.mixing.dist(CI.est.res)
}
\arguments{
\item{CI.est.res}{Result of \code{\link{mcleod.estimate.CI}}, run with parameter \code{compute_CI_curves} set to \code{T}}
}
\value{
a data frame with three columns:
#' \itemize{
\item{\code{log.odds}}{ - specificying points in the suppor of the mixing distribution, in terms of \eqn{P/log(1-P)}.}
\item{\code{Lower}}{ - Specifying a lower bound on the CDF, for the corresponding value of \code{log.odds}.}
\item{\code{Upper}}{ - Specifying an upper bound on the CDF, for the corresponding value of \code{log.odds}.}
}
;
}
\description{
Pointwise confidence intervals computed in confidence levels set by \code{mcleod.estimate.CI}.
}
\details{
Resolution of \code{log.odds} given by parameter \code{theta_vec} in \code{mcleod.CI.estimation.parameters} (used to define object passed to \code{mcleod.estimate.CI}).
}
\examples{
 # For full description of package model and workflow,
 # including this function, Type browseVignettes(package = 'mcleod') 
 # in the R console and check the package vignette
 
 ################################
 ## Example 
 ################################
 #Generate data
 
 n = 500
 N = rep(20,n) 
 set.seed(1)
 p = inv.log.odds(rnorm(n,-2,0.5)+3*rbinom(n,1,0.3))
 X = rbinom(n = n,size = N,prob = p)
 
 # Compute point-wise CIs for the mixing distribution
 CI.est.res = mcleod.estimate.CI(X = X, N = N,
         CI_param = mcleod.CI.estimation.parameters(rho.set.value = 0.25))
 
 # Plot the pointwise CIs
 plot.mcleod.CI(mcleod.CI.obj = CI.est.res)
 
 # Retreive CIs for each log-odds value as a data table
 dt_CIs = mcleod.get.CIs.mixing.dist(CI.est.res)
 dt_CIs
 
}
