% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{mcleod.get.posterior.mixing.dist}
\alias{mcleod.get.posterior.mixing.dist}
\title{Retrieve the estimated posterior mixing distribution, both density and CDF}
\usage{
mcleod.get.posterior.mixing.dist(
  res,
  aggregate_by = mean,
  specific_iter = NULL
)
}
\arguments{
\item{res}{object resulting from function \code{\link{mcleod}}}

\item{aggregate_by}{function to aggregate by, such as \code{\link{mean}} or \code{\link{median}} (used for computing the posterior mean / median, respectively, at each point of the support of the mixing distribution). May also be a user supplied function.}

\item{specific_iter}{Default value is \code{NULL}, meaning the function returns estimates obtained via posterior mean/ median across all gibbs samples. If the user picks a single number, this iteration of the gibbs sample is returned.}
}
\value{
A list with the following fields:
\itemize{
\item{CDF}{ - if \code{specific_iter} is set to \code{NULL}, returns an \code{R} \code{\link{approxfun}} function, giving an estimate (pointwise posterior mean/median) of the CDF of the mixing distribution. If a specific iteration is requested, will return the CDF for mixing distribution, found in the \code{specific_iter} iteration of the MCMC sampler.}
\item{density}{ - if \code{specific_iter} is set to \code{NULL}, returns an \code{R} \code{\link{approxfun}} function, giving an estimate (pointwise posterior mean/median) of the density of the mixing distribution. If a specific iteration is requested, will return the density for mixing distribution, found in the \code{specific_iter} iteration of the MCMC sampler.}
\item{a.vec}{ - Parameter vector \eqn{\vec{a}}, used to define the support of the piece-wise constant mixing distribution. See full definition in the package vignette.}
\item{pi_smp}{ - Parameter vector \eqn{\vec{\pi}}, detailing the estiamted probability mass in each segment of \eqn{\vec{a}}. Computed via pointwise posterior mean/median across MCMC samples. See full definition in the package vignette. If a specific iteration is requested, will return the value of \eqn{\vec{\pi}} sampled in that specific iteration.}
}
}
\description{
The function computes the point-wise posterior mean/median of the mixing distribution, using the MCMC samples computed by \code{\link{mcleod}}. The user may also ask for the mixing distribution (CDF,density and \eqn{\vec{\pi}}) from a specific iteration of the Gibbs sampler.
}
\details{
Posterior mean/ median computed in a pointwise manner, across points of the support, in order to obtain an estimate of the mixing distribution (which is a function). Burn-in samples (specified by the function \code{\link{mcleod}}) are removed from the MCMC samples when computing estiamtes.
}
\examples{
see package vignette
}
