% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interval_Censoring_Regression.R
\name{mcleod.interval.censoring.density.estimation}
\alias{mcleod.interval.censoring.density.estimation}
\title{Title}
\usage{
mcleod.interval.censoring.density.estimation(lu.mat, L = 8,
  a.rng = c(0, 1100), n.gbbs = 200, n.gbbs.burnin = 100)
}
\arguments{
\item{lu.mat}{}

\item{L}{}

\item{a.rng}{}

\item{n.gbbs}{}

\item{n.gbbs.burnin}{}
}
\description{
Title
}
