% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med_names}
\alias{med_names}
\title{Medication Names}
\format{a \code{data.frame} with 1 variable
\describe{
  \item{GENERIC_MED_NAME}{(chr) Generic medication names that occur in the PRESCRIBING table}
}}
\source{
med_names.csv
}
\usage{
med_names
}
\description{
This table contains the Medication names Lisa McLeod's group are interested in .
}
\keyword{datasets}
