% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med_names2}
\alias{med_names2}
\title{Medication Names 2}
\format{a \code{data.frame} with 1 variable
\describe{
  \item{GENERIC_MED_NAME}{(chr) Generic medication names that occur in the PRESCRIBING table}
}}
\source{
med-names2.csv
}
\usage{
med_names2
}
\description{
This table contains the second list medication names Lisa McLeod's group are interested in.
}
\keyword{datasets}
