% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med_names3}
\alias{med_names3}
\title{Medication Names 3}
\format{a \code{data.frame} with 1 variable
\describe{
  \item{GENERIC_MED_NAME}{(chr) Generic medication names that occur in the PRESCRIBING table}
}}
\source{
med-names3.csv
}
\usage{
med_names3
}
\description{
This table contains the third list medication names Lisa McLeod's group are interested in.
}
\keyword{datasets}
