% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{plot.mcleod.CI}
\alias{plot.mcleod.CI}
\title{Title}
\usage{
\method{plot}{mcleod.CI}(
  mcleod.CI.obj,
  X_axis_as_Prob = T,
  add_CI_curves_on_top_of_plot = F,
  point_estimate_color = "red",
  CI_curves_color = "black",
  title = ""
)
}
\arguments{
\item{mcleod.CI.obj}{}

\item{X_axis_as_Prob}{}

\item{add_CI_curves_on_top_of_plot}{}

\item{point_estimate_color}{}

\item{CI_curves_color}{}

\item{title}{}
}
\value{

}
\description{
Title
}
