% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interval_Censoring_Regression.R
\name{plot.posterior.interval.censoring}
\alias{plot.posterior.interval.censoring}
\title{Title}
\usage{
\method{plot}{posterior.interval.censoring}(mcleod.ice.obj)
}
\arguments{
\item{mcleod.ice.obj}{}
}
\description{
Title
}
