% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surg2rec.R
\docType{package}
\name{surg2rec}
\alias{surg2rec}
\alias{surg2rec-package}
\title{surg2rec: The goal of surg2rec is to generate the tables requested by Lisa McLeod.}
\description{
The surg2rec package contains the function, .RMD files, and data files necessary for generating tables from CHORDS data for Surgery to Recovery project.
}
