% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{compute_CI_curves_function}
\alias{compute_CI_curves_function}
\title{Function performs efficient computation of pointwise CIs, for a given confidence level}
\usage{
compute_CI_curves_function(
  bank_original,
  res_mcleod_data,
  rho_calibration_obj,
  median_curve,
  alpha.one.sided,
  verbose
)
}
\arguments{
\item{bank_original}{}

\item{res_mcleod_data}{}

\item{rho_calibration_obj}{}

\item{median_curve}{}

\item{alpha.one.sided}{}

\item{verbose}{}
}
\value{
a list with the following entries: \code{q_star_LE} - ; \code{q_star_GE} - ; \code{bank} -
}
\description{
Function performs efficient computation of pointwise CIs, for a given confidence level
}
\keyword{internal}
