% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{compute_medians_curve}
\alias{compute_medians_curve}
\title{Title}
\usage{
compute_medians_curve(mcleod_for_data, list_ind = NA)
}
\arguments{
\item{mcleod_for_data}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
