% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{compute_P_values_over_grid_function}
\alias{compute_P_values_over_grid_function}
\title{Title}
\usage{
compute_P_values_over_grid_function(
  bank_original,
  rho_calibration_obj,
  res_mcleod_data,
  median_curve,
  alpha.one.sided,
  verbose = F
)
}
\arguments{
\item{bank_original}{}

\item{rho_calibration_obj}{}

\item{res_mcleod_data}{}

\item{median_curve}{}

\item{alpha.one.sided}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
