% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{generate_P_k_i_matrix_cache}
\alias{generate_P_k_i_matrix_cache}
\title{Title}
\usage{
generate_P_k_i_matrix_cache(n.vec, a.max, prior_param, comp_param)
}
\arguments{
\item{n.vec}{}

\item{a.max}{}

\item{prior_param}{}

\item{comp_param}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
