% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod-package.R
\docType{package}
\name{mcleod-package}
\alias{mcleod-package}
\alias{_PACKAGE}
\alias{mcleod-pkg}
\title{mcleod: Monte Carlo methods for Estimating Latent OverDispersion of binomial and Poisson random variables.}
\description{
The package contains several heirarchical Bayesian methods for estimating the mixing distribution of Binomial and Poisson samples. Methods include regularized estimators of the CDF of the mixing distribution, confidence intervals for the CDF of the mixing distribution, and binomial and Poisson regression model where the random intercept term may have a general distribution.
}
\keyword{internal}
