% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.CI.deconv.bank.get_median_curves_for_worst_case_hypothesis_at_point}
\alias{mcleod.CI.deconv.bank.get_median_curves_for_worst_case_hypothesis_at_point}
\title{Title}
\usage{
mcleod.CI.deconv.bank.get_median_curves_for_worst_case_hypothesis_at_point(
  bank,
  ind_theta,
  ind_q,
  a_index,
  nr.perms,
  is_GE = T,
  do_serial = F
)
}
\arguments{
\item{bank}{}

\item{ind_theta}{}

\item{ind_q}{}

\item{a_index}{}

\item{nr.perms}{}

\item{is_GE}{}

\item{do_serial}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
