% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_CDF_Procedure.R
\name{mcleod.CI.estimation.parameters}
\alias{mcleod.CI.estimation.parameters}
\title{Title}
\usage{
mcleod.CI.estimation.parameters(Nr.reps.for.each.n = 1,
  nr.cores = detectCores() - 1, epsilon.nr.gridpoints = 2,
  fraction.of.points.computed = 1)
}
\arguments{
\item{Nr.reps.for.each.n}{}

\item{nr.cores}{}

\item{epsilon.nr.gridpoints}{}

\item{fraction.of.points.computed}{}
}
\description{
Title
}
