% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.CI.lower_bound_PV_for_worst_case}
\alias{mcleod.CI.lower_bound_PV_for_worst_case}
\title{Title}
\usage{
mcleod.CI.lower_bound_PV_for_worst_case(bank, ind_q, CDF_value, is_GE = F)
}
\arguments{
\item{bank}{}

\item{ind_q}{}

\item{CDF_value}{}

\item{is_GE}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
