% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.CI.rho.calibration.constructor}
\alias{mcleod.CI.rho.calibration.constructor}
\title{Title}
\usage{
mcleod.CI.rho.calibration.constructor(
  bank_original,
  res_mcleod_holdout,
  CDF_holdout,
  alpha.one.sided,
  nr.perm = 200,
  possible_rhos = c(0.1, 0.2, 0.3, 0.4, 0.5),
  q_for_rho_optimization = c(0.2, 0.4, 0.6, 0.8),
  theta_for_rho_optimization = NULL,
  verbose = F
)
}
\arguments{
\item{bank_original}{}

\item{res_mcleod_holdout}{}

\item{CDF_holdout}{}

\item{alpha.one.sided}{}

\item{nr.perm}{}

\item{possible_rhos}{}

\item{q_for_rho_optimization}{}

\item{theta_for_rho_optimization}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
