% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.CI.rho.calibration.constructor}
\alias{mcleod.CI.rho.calibration.constructor}
\title{Function performs calibration for the value of rho used for testing (unless the user asked for a specific rho)}
\usage{
mcleod.CI.rho.calibration.constructor(
  bank_original,
  res_mcleod_holdout,
  CDF_holdout,
  alpha.one.sided,
  nr.perm = 200,
  possible_rhos = c(0.1, 0.2, 0.3, 0.4, 0.5),
  q_for_rho_optimization = c(0.2, 0.4, 0.6, 0.8),
  theta_for_rho_optimization = NULL,
  verbose = F
)
}
\arguments{
\item{bank_original}{Cache object to be updated. Also contains CI_params containing required parameters and definitions.}

\item{res_mcleod_holdout}{mcleod result for the holdout data}

\item{CDF_holdout}{point estimates for the CDF (posterior median), obtained for the holdout data}

\item{alpha.one.sided}{significance level for one sided test}

\item{nr.perm}{number of bootstrap samples when testing}

\item{possible_rhos}{an ordered vector of the possible rhos to be considered}

\item{q_for_rho_optimization}{an ordered vector of CDF values, used to select the thetas for which optimization is performed (by looking at the thetas corresponding to these CDF values in CDF_holdout)}

\item{theta_for_rho_optimization}{if specified, thetas for which calibration is performed will be taken from this vector, instead of the ones defined by the CDF values given by \code{q_for_rho_optimization}}

\item{verbose}{should the function print messages}
}
\value{
an object of type 'mcleod.CI.obj.rho' with the following fields
\itemize{
\item{rho_approx_fun_LE}{ - a function, giving the value of rho for LE hypotheses, for each value of theta* tested. If user asked for a fixed value of rho, this function returns a constant.}
\item{rho_approx_fun_GE }{ -  - a function, giving the value of rho for GE hypotheses, for each value of theta* tested. If user asked for a fixed value of rho, this function returns a constant.}
\item{bank }{ - updated cache to be inserted back in the calling function}
\item{rho_approx_fun_LE_specific_non_smoothed }{ - a function, giving the value of rho for LE hypotheses, for each value of theta* tested, for case the user used the parameter \code{rho.calibration.method='specific'}. This function is computed anyway, and if the user picked 'specific', it is copied to \code{rho_approx_fun_LE} and used in actual testing.}
\item{rho_approx_fun_GE_specific_non_smoothed }{ - same as \code{rho_approx_fun_LE_specific_non_smoothed}, only for GE hypotheses.}
\item{rho_approx_fun_LE_specific_smoothed }{ - A kernel smoothed version of \code{rho_approx_fun_LE_specific_non_smoothed}, currently not used in computations.}
\item{rho_approx_fun_GE_specific_smoothed }{ - Same as \code{rho_approx_fun_LE_specific_smoothed}, only for GE hypotheses.}
\item{curve_LE_q_by_theta_and_rho }{ - a matrix, giving for each theta point on which the CI is optimized (row), and each value of rho (col), the upper bound (upper edge of CI) for the CDF value.}
\item{curve_GE_q_by_theta_and_rho }{ - a matrix, giving for each theta point on which the CI is optimized (row), and each value of rho (col), the lower bound (lower edge of CI) for the CDF value.}
\item{distances }{ - Available only if user selected \code{rho.calibration.method} to be 'max' or 'sum'. Gives the element-wise subtraction of \code{curve_GE_q_by_theta_and_rho} from \code{curve_LE_q_by_theta_and_rho}, thus giving the length of confidence intervals for each theta point and value of rho.}
\item{scores}{ - Available only if user selected \code{rho.calibration.method} to be 'max' or 'sum'. Aggregation of \code{distances} by column-wise max or sum.}
\item{selected_rho}{ - Available only if user selected \code{rho.calibration.method} to be 'max' or 'sum'. value of rho corresponding to the minimum entry of \code{scores}.}
}
}
\description{
Calibration results returned as an object of a specific type
}
\keyword{internal}
