% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{mcleod.computational.parameters}
\alias{mcleod.computational.parameters}
\title{Title}
\usage{
mcleod.computational.parameters(nr.gibbs = 500, nr.gibbs.burnin = 250,
  integration_step_size = 0.01, Fast.Gamma.Used = F,
  Fast.Gamma.Bank.Size = 1000L)
}
\arguments{
\item{nr.gibbs}{}

\item{nr.gibbs.burnin}{}

\item{integration_step_size}{}

\item{Fast.Gamma.Used}{}

\item{Fast.Gamma.Bank.Size}{}
}
\description{
Title
}
