% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{mcleod.covariates.estimation.parameters}
\alias{mcleod.covariates.estimation.parameters}
\title{An object of type 'mcleod.covariates.estimation.parameters.obj' defining computational and statistical parameters, when estimating a random effect model with covariates.}
\usage{
mcleod.covariates.estimation.parameters(
  proposal_sd = c(0.05),
  beta_prior_sd = c(1),
  beta_init = c(0),
  Manual_Prior_Values = NULL,
  Manual_Prior_Probs = NULL,
  do_P_k_i_hashing = F,
  P_k_i_hashing_resolution_by_theta = 1e-04
)
}
\arguments{
\item{proposal_sd}{vector of standard deviations for the proposal distribution, for the values of delta. Proposals are generated by adding a random value from N(0,diag(proposal_sd)) to the current MCMC value of delta.}

\item{beta_prior_sd}{vector of standard deviations for the prior. The prior for delta is assumed to be N(0,diag(beta_prior_sd)). Setting an entry to a negative value (e.g. -1) will use an uninformative prior (constant) for that component of delta.}

\item{beta_init}{Initial values for beta. default value is zero, however non-cold-start values can be used, e.g. by using values from a binomial, or normal random-intercept binomial regression.}

\item{Manual_Prior_Values}{When using a predefined prior for delta, a vector of sorted values defining the paritioning points for a piece-wise constant prior for delta.}

\item{Manual_Prior_Probs}{A vector with length = length(Manual_Prior_Values)-1, giving the probablities for the different segments defined by  \code{Manual_Prior_Values}. Within each segment, it is assumed the prior is uniformly distributed.}

\item{do_P_k_i_hashing}{A computational workaround for faster computation of bayesian probabilites. See additional details in the appendix describing the computational method, in the dissertation of Brill (2022)}

\item{P_k_i_hashing_resolution_by_theta}{Resolution parameter for the computational workaround activated by do_P_k_i_hashing}
}
\value{
an object of type 'mcleod.covariates.estimation.parameters.obj'
}
\description{
An object of type 'mcleod.covariates.estimation.parameters.obj' defining computational and statistical parameters, when estimating a random effect model with covariates.
}
\details{
For binomial response, the model with covariates is log(p/(1-p)) = gamma + delta^T Z, where gamma is a random deviate from a general density estimated from the data, delta is a vector of slopes, and Z is a vector of covariates. For poisson errors, the mode equation is log(lambda) = gamma + delta^T Z
}
\examples{
See the package vignette
}
