% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{mcleod.covariates.estimation.parameters}
\alias{mcleod.covariates.estimation.parameters}
\title{Title}
\usage{
mcleod.covariates.estimation.parameters(proposal_sd = c(0.05),
  beta_prior_sd = c(5), beta_init = c(0), Manual_Prior_Values = NULL,
  Manual_Prior_Probs = NULL)
}
\arguments{
\item{proposal_sd}{}

\item{beta_prior_sd}{}

\item{beta_init}{}

\item{Manual_Prior_Values}{}

\item{Manual_Prior_Probs}{}
}
\description{
Title
}
