% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_CDF_Procedure.R
\name{mcleod.estimate.CI}
\alias{mcleod.estimate.CI}
\title{Title}
\usage{
mcleod.estimate.CI(x.vec, n.vec, q_grid = seq(0.1, 0.9, 0.1),
  a.max = c(3), comp_param = mcleod.computational.parameters(nr.gibbs =
  200, nr.gibbs.burnin = 100),
  prior_param = mcleod.prior.parameters(prior.type =
  MCLEOD.PRIOR.TYPE.TWO.LAYER.DIRICHLET, Two.Layer.Dirichlet.Intervals =
  64, Two.Layer.Dirichlet.Nodes.in.First.Layer = 8),
  CI.estimation.parameters = mcleod.CI.estimation.parameters(),
  verbose = T)
}
\arguments{
\item{x.vec}{}

\item{n.vec}{}

\item{q_grid}{}

\item{a.max}{}

\item{comp_param}{}

\item{prior_param}{}

\item{CI.estimation.parameters}{}

\item{verbose}{}
}
\description{
Title
}
