% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.estimate.CI}
\alias{mcleod.estimate.CI}
\title{Compute point-wise Confidence Intervals for the Quantiles and Percentiles of the mixing distribution of binomial samples.}
\usage{
mcleod.estimate.CI(
  X,
  N,
  CI_param = mcleod.CI.estimation.parameters(),
  ratio_holdout = 0.1,
  compute_P_values_over_grid = F,
  compute_CI_curves = T,
  verbose = T,
  Use_Existing_Permutations_From_Object = NULL
)
}
\arguments{
\item{X}{Number of successful draws per observation}

\item{N}{Number of draws per observation}

\item{CI_param}{Result of function \code{\link{mcleod.CI.estimation.parameters}}, setting statistical and computational parameters. If you are unsure about the parameters, best to use the default option.}

\item{ratio_holdout}{When estimating rho adaptively from the data, the ratio of data samples used for the calibration procedure for rho. Values such as 0.1 and 0.2 are plausible.}

\item{compute_P_values_over_grid}{Shoudl}

\item{compute_CI_curves}{}

\item{verbose}{Should messages be printed to screen.}

\item{Use_Existing_Permutations_From_Object}{An existing object generated using mcleod.estimate.CI. The cache of CDF samples (used for computing PVs) is instantiated using the CDF samples available in this object. This helps speed up computations considerably. NOTE: It is up to the user to verify the the object uses the same CI parameters and N vec. If the used CDF samples (from WC hypothesis tests) were computed CI.params and N.vec different from the ones in the data, the results might be meaningless!}
}
\value{
A list set as class 'mcleod.CI.obj' with the following entries:
\itemize{
\item{pvalues_grid}{ - available if compute_P_values_over_grid is set to True. Contains a list with entries as detailed below. }
\item{computed_curves}{ - available if compute_CI_curves is set to True. Contains a list with entries as detailed below.}
\item{bank}{ - results of the internal function \code{\link{mcleod.CI.deconv.bank.constructor}}. This object serves as a cache storing estimated CDF of the mixing distribution, from bootstrap data samples used for estimating P.values. See more information in the package vignette. When running the function on similar inputs (having the same value of N.vec and the same CI params) this object can used to speed up PV computations, see description for above parameter \code{Use_Existing_Permutations_From_Object} }
\item{rho_calibration_obj}{ - result of internal function \code{\link{mcleod.CI.rho.calibration.constructor}}. Contains information on how rho was set adaptively for the data.}
\item{n_holdout}{ - When splitting the data, in order to choose rho adaptively, this is the number of samples taken from the full data and used for calibration.}
\item{X_rho}{ - number of successfull draws, for the part of the data used for calibrating rho. A subset of X}
\item{N_rho}{ - number of draws, for the part of the data used for calibrating rho. A subset of N, corresponding to \code{X_rho}}
\item{X_test}{ - number of successfull draws, for the part of the data used for testing. The remainder of X, after excluding \code{X_rho}}
\item{N_test}{ - number of draws, for the part of the data used for testing. The remainder of N, after excluding \code{N_rho}}
\item{res_mcleod_holdout}{ - result of \code{\link{mcleod}} run on the holdout data, used for calibrating rho.}
\item{CDF_holdout}{ - Values of the estimated CDF for mixture distribution in the holdout data (used for calibrating rho). The estimator for the mixing distribution is the pointwise (by theta) median of the posterior distribution. The theta values (X-axis) for the CDF are given by res_mcleod_holdout$parameters_list$a.vec}
\item{res_mcleod_data}{ - result of \code{\link{mcleod}} run on the test data. }
\item{CDF_data}{ - Values of the estimated CDF for mixture distribution in the test data. The estimator for the mixing distribution is the pointwise (by theta) median of the posterior distribution. The theta values (X-axis) for the CDF are given by res_mcleod_data$parameters_list$a.vec}
\item{ind_selected_for_test}{ - When splitting the data (due to using part of the data for calibrating rho), these are the indices (in terms of the input, X and N) of the samples used for testing.}
}

The list given by \code{computed_curves} contains the following entries:
\itemize{
\item{q_star_LE}{ - The upper end of pointwise CDFs for the mixing distribution, by value of theta. Values of theta are given in the vector entry names.}
\item{q_star_GE}{ - The lower end of pointwise CDFs for the mixing distribution, by value of theta. Values of theta are given in the vector entrynames.}
}

The list given by \code{pvalues_grid}
\itemize{
\item{GE.pval.grid}{ - Contains the PVs for GE type hypotheses. Coordinate by q and theta are given by column and row names.}
\item{LE.pval.grid}{ - Contains the PVs for LE type hypotheses. Coordinate by q and theta are given by column and row names.}
}
}
\description{
The assumed model is of the form \eqn{Xi ~ bin(Ni,Pi)}, where \eqn{i} is an index in the range 1,...,n. The function constructs pointwise confidence intervals for the mixing distribution, i.e., the distribution of Pi's. The statistical method is explained in the package vignette.
}
\examples{
See package vignette
}
