% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.estimate.CI.single.q}
\alias{mcleod.estimate.CI.single.q}
\title{Title}
\usage{
mcleod.estimate.CI.single.q(
  X,
  N,
  q,
  CI_param = mcleod.CI.estimation.parameters(),
  ratio_holdout = 0.1,
  verbose = F
)
}
\arguments{
\item{X}{}

\item{N}{}

\item{q}{}

\item{CI_param}{}

\item{ratio_holdout}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
