% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{mcleod.get.CIs.mixing.dist}
\alias{mcleod.get.CIs.mixing.dist}
\title{Extract pointwise confidence intervals for the CDF of the mixing distribution, from result of function \code{mcleod.estimate.CI}}
\usage{
mcleod.get.CIs.mixing.dist(CI.est.res)
}
\arguments{
\item{CI.est.res}{Result of \code{\link{mcleod.estimate.CI}}, run with parameter \code{compute_CI_curves} set to \code{T}}
}
\value{
a data frame with three columns:
#' \itemize{
\item{\code{log.odds}}{ - specificying points in the suppor of the mixing distribution, in terms of \eqn{P/log(1-P)}.}
\item{\code{Lower}}{ - Specifying a lower bound on the CDF, for the corresponding value of \code{log.odds}.}
\item{\code{Upper}}{ - Specifying an upper bound on the CDF, for the corresponding value of \code{log.odds}.}
}
;
}
\description{
Pointwise confidence intervals computed in confidence levels set by \code{mcleod.estimate.CI}.
}
\details{
Resolution of \code{log.odds} given by parameter \code{theta_vec} in \code{mcleod.CI.estimation.parameters} (used to define object passed to \code{mcleod.estimate.CI}).
}
\examples{
see package vignette
}
