% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{mcleod.posterior.estimates.random.effect}
\alias{mcleod.posterior.estimates.random.effect}
\title{Compute posterior estimates for \eqn{P_i} and \eqn{\lambda_i}}
\usage{
mcleod.posterior.estimates.random.effect(
  X,
  N,
  mcleod_res,
  covariates = NULL,
  method = "mean",
  offset_vec = rep(0, length(X))
)
}
\arguments{
\item{X}{A vector of integers, with each entry corresponding to an observations. For binomial data, the number of successful draws for each observations. For Poisson data, the number of counts for each observation.}

\item{N}{For binomial data, a vector giving the number of draws for each observations. Entries correspond to entries in \code{X}. For Poisson data, set to \code{NULL}.}

\item{mcleod_res}{Model fitted using \code{\link{mcleod}}, holding an estimate of the mixing distribution.}

\item{covariates}{A matrix of covariates, with rows corresponding to the entries of X, and columns corresponding to the covariates used when fitting \code{mcleod_res}. If the mixing distribution was estimated without covariates, covariates cannot be used here, set the value of this parameter to \code{NULL}.}

\item{method}{Type of estimator for \eqn{P_i} \ \eqn{\lambda_i} \ \eqn{\gamma_i}. Use \code{'mean'} for posterior mean, and \code{'mode'} for posterior mode.}

\item{offset_vec}{A vector of offset values, for the linear predictor term. See full definition in the function vignette.}
}
\value{
Vector of estimates for \eqn{P_i} (for binomial data) , \eqn{\lambda_i} (for Poisson data), or \eqn{\gamma_i} (observation's random intercept, for data with covariates). Entries correspond to entries of \code{X}
}
\description{
Function computes posterior estimates for \eqn{P_i} and \eqn{\lambda_i} for observations, based on observed counts and fitted \code{\link{mcleod}} model. The function receives as input the fitted mcleod model, as well as the observables for the data: the observed counts (X, for both Poisson and Binomial observations), number of draws (relevant for binomial observations), covariates and offset term. The type of data supplied (Binomial/Poisson, with or without covariates) should be identical to the type of data used to fit the \code{\link{mcleod}} model supplied.
}
\details{
For data with covariates, the returned values are estimates for \eqn{\gamma_i}, the observation specific random intercept term.
}
\examples{
See package vignette
}
