% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{mcleod.predictive.interval}
\alias{mcleod.predictive.interval}
\title{Compute prediction intervals for out-of-sample observations}
\usage{
mcleod.predictive.interval(
  N,
  mcleod_res,
  covariates = NULL,
  offset_vec = NULL,
  Interval.Coverage = 0.95,
  Nr_Simulated_Values = 1000
)
}
\arguments{
\item{N}{For binomial data, a vector specifying the number of draws for each sample. For Poisson data, set to \code{NULL}.}

\item{mcleod_res}{Model fitted using \code{\link{mcleod}}.}

\item{covariates}{Matrix of covariates for the out-of-sample observations with rows corresponding to observations, and columns corresponding to the variables used to fit \code{mcleod_res}.}

\item{offset_vec}{A vector of offset values for the linear predictor. For binomial data, must have length equal to \code{length(N)} and \code{nrow(covariates)}. For Poisson data, must have length equal to  \code{nrow(covariates)}.}

\item{Interval.Coverage}{Coverage probability for the prediction interval.}

\item{Nr_Simulated_Values}{Number of simulation replicates, used for computed prediction intervals.}
}
\value{
A list with two vectors named \code{Lower} and \code{Upper}, giving the lower and upper ends of prediction intervals, respectively. for the out-of-sample observations defined by \code{N},\code{covariates} and \code{offset_vec}.
}
\description{
The function computes prediction intervals, for out-of-sample observations (i.e., not part of the training set), based on a \code{\link{mcleod}} model. The type of observations for which predictive intervals are computed (Binomial/Poisson, with or without covariates), is determined by the type of model passed as \code{mcleod_res}. For binomial samples, the function receives \code{N}, a vector specifying the number of draws for each observations. The function then returns predictive intervals for each observations, see section 'return' below. For Poisson samples without covariates, all samples are identically distributed, and the function returns a single prediction interval. If the model supplied under \code{mcleod_res} uses covariates, the function must also be supplied covariates for each observation. Covariates are inserted under \code{covariates}, as matrix with rows corresponding to observations, and columns corresponding to the variables used to fit \code{mcleod_res}. For binomial samples, the number of rows of \code{covariates} must be identical to the length of \code{N}. For Poisson samples, if covariates are given, the function will compute predictive intervals for multiple Poisson samples, corresponding to the rows of \code{covariates}.
}
\details{
The function computes prediction intervals by simulating counts using the fitted model, i.e., sampling values from the random effect from the fitted prior, adjusting for covariates and offset terms, and using Binomial/Poisson sampling to generate counts. A higher number of simulation replicates, determined by \code{Nr_Simulated_Values}, means more accurate prediction intervals. For most applications, 1000 replicates are satisfactory. However, if tail probabilities are needed (e.g. 0.9999 prediction interval), a larger value for \code{Nr_Simulated_Values} needs to be set.
}
\examples{
See package vignette
}
