% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{mcleod.prior.parameters}
\alias{mcleod.prior.parameters}
\title{Title}
\usage{

  mcleod.prior.parameters(prior.type = MCLEOD.PRIOR.TYPE.BETA.HEIRARCHICAL,
  Beta.Heirarchical.Levels = 6, Two.Layer.Dirichlet.Intervals = 64,
  Two.Layer.Dirichlet.Nodes.in.First.Layer = 8)
}
\arguments{
\item{prior.type}{}

\item{Beta.Heirarchical.Levels}{}

\item{Two.Layer.Dirichlet.Intervals}{}

\item{Two.Layer.Dirichlet.Nodes.in.First.Layer}{}
}
\description{
Title
}
