% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{mcleod}
\alias{mcleod}
\title{Title}
\usage{
mcleod(x.smp, n.smp, a.limits = c(-4, 4),
  Noise_Type = MCLEOD.BINOMIAL.ERRORS, covariates = NULL,
  prior_parameters = NULL, computational_parameters = NULL,
  covariates_estimation_parameters = NULL, input_P_k_i = NULL,
  exact.numeric.integration = TRUE)
}
\arguments{
\item{x.smp}{}

\item{n.smp}{}

\item{a.limits}{}

\item{Noise_Type}{}

\item{covariates}{}

\item{prior_parameters}{}

\item{computational_parameters}{}

\item{covariates_estimation_parameters}{}

\item{input_P_k_i}{}
}
\value{
names(res$parameters_list)
\link{1} "a.vec"                            "nr.gibbs"
\link{3} "nr.gibbs.burnin"                  "prior_parameters"
\link{5} "computational_parameters"         "covariates_estimation_parameters"
\link{7} "covariates"                       "x.smp"
\link{9} "n.smp"                            "Noise_Type"
\link{11} "covariates_given"
names(res$additional$original_stat_res)
\link{1} "p_k_i"             "n_smp"             "pi_smp"            "beta_smp"          "beta_suggestion"
\link{6} "proposal_approved" "elapsed_secs"      "ll_proposal"       "ll_current"
}
\description{
Title
}
