% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mixture_distribution.R
\name{plot.mcleod.CI}
\alias{plot.mcleod.CI}
\title{Plot the pointwise Confidence Intervals obtained from running mcleod.estimate.CI}
\usage{
\method{plot}{mcleod.CI}(
  mcleod.CI.obj,
  X_axis_as_Prob = T,
  add_CI_curves_on_top_of_plot = F,
  point_estimate_color = "red",
  CI_curves_color = "black",
  title = "",
  type = "p",
  lwd = 1,
  lwd_CI_curves = 1,
  lty = 1,
  lty_CI_curves = 1,
  override_xlab = NULL,
  plot_Point_Estimate = T
)
}
\arguments{
\item{mcleod.CI.obj}{An object returned from \code{\link{mcleod.estimate.CI}}, with parameter \code{compute_CI_curves} set to \code{True}.}

\item{X_axis_as_Prob}{Should the X axis for the mixing distribution be presented as Probability (True, default) or as log-odds (parameter named theta, False)}

\item{add_CI_curves_on_top_of_plot}{Should the CI curves be plotted on top of an existing plot. This option is usefull when plotting multiple CIs for the same data.}

\item{point_estimate_color}{Color for the posterior median of the mixing distribution. Medians are taken point-wise across the support.}

\item{CI_curves_color}{Color for the curves showing the pointwise confidence intervals.}

\item{title}{Title for the plot}

\item{type}{overrides the plot type use 'l' or 'p'}

\item{lwd}{line width for CDF estimate}

\item{lwd_CI_curves}{= line width for pointwise CI curves}

\item{lty}{= line type for CDF estimate}

\item{lty_CI_curves}{= line type for pointwise CI curves}

\item{override_xlab}{used to override the plot x label}
}
\value{
NONE
}
\description{
Plot the pointwise Confidence Intervals obtained from running mcleod.estimate.CI
}
\examples{
 # For full description of package model and workflow,
 # including this function, Type browseVignettes(package = 'mcleod') 
 # in the R console and check the package vignette
 
 ################################
 ## Example 
 ################################
 #Generate data
 
 n = 500
 N = rep(20,n) 
 set.seed(1)
 p = inv.log.odds(rnorm(n,-2,0.5)+3*rbinom(n,1,0.3))
 X = rbinom(n = n,size = N,prob = p)
 
 # Compute point-wise CIs for the mixing distribution
 CI.est.res = mcleod.estimate.CI(X = X, N = N,
         CI_param = mcleod.CI.estimation.parameters(rho.set.value = 0.25))
 
 # Plot the pointwise CIs
 plot.mcleod.CI(mcleod.CI.obj = CI.est.res)
 
 # Retreive CIs for each log-odds value as a data table
 dt_CIs = mcleod.get.CIs.mixing.dist(CI.est.res)
 dt_CIs
 
}
