% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_CDF_Procedure.R
\name{plot.mcleod.CI}
\alias{plot.mcleod.CI}
\title{Title}
\usage{
\method{plot}{mcleod.CI}(mcleod.CI.obj, sig.level = c(0.05, 0.1))
}
\arguments{
\item{mcleod.CI.obj}{}

\item{sig.level}{}
}
\description{
Title
}
