% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{plot.posterior}
\alias{plot.posterior}
\title{Title}
\usage{
\method{plot}{posterior}(mcleod.obj, plot_only_point_estimate = F)
}
\arguments{
\item{mcleod.obj}{}

\item{plot_only_point_estimate}{}
}
\value{

}
\description{
Title
}
