% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcleod_main_function.R
\name{results.covariate.coefficients.posterior}
\alias{results.covariate.coefficients.posterior}
\title{Title}
\usage{
results.covariate.coefficients.posterior(mcleod.obj, plot.posterior = T,
  plot.MH.proposal.by.iteration = F)
}
\arguments{
\item{mcleod.obj}{}

\item{plot.posterior}{}

\item{plot.MH.proposal.by.iteration}{}
}
\description{
Title
}
